+++ clean.sh
@@ -8,7 +8,6 @@
 
 # Color codes for output
 GREEN='\033[0;32m'
-RED='\033[0;31m'
 YELLOW='\033[0;33m'
 BLUE='\033[0;34m'
 CYAN='\033[0;36m'